agents/adk-agent/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
pandas
numpy
scipy
python-multipart
httpx
python-dotenv